fix: removed crash when message deleted error

+++ src/commands/help.mts
@@ -1,9 +1,9 @@
-import { ActionRowBuilder, ButtonBuilder, ButtonStyle, ChatInputCommandInteraction, Collection, Message, MessageFlags, ModalBuilder, SlashCommandBuilder, TextChannel, TextInputBuilder, TextInputStyle, ThreadAutoArchiveDuration, type Snowflake } from "discord.js";
+import { ActionRowBuilder, ButtonBuilder, ButtonStyle, ChatInputCommandInteraction, Collection, EmbedBuilder, Message, MessageFlags, ModalBuilder, SlashCommandBuilder, TextChannel, TextInputBuilder, TextInputStyle, ThreadAutoArchiveDuration, type Snowflake } from "discord.js";
 import { User } from "../class/User.mts";
 import { Configuration } from "../class/Configuration.mts";
 import { Log } from "../class/Log.mts";
 import { Bot } from "../class/Bot.mts";
-import { replaceMentionsWithUsernames, replaceUsernamesWithMentions, splitFromJumpLines } from "../lib/helpers.mts";
+import { threadStillExists, replaceMentionsWithUsernames, replaceUsernamesWithMentions, splitFromJumpLines } from "../lib/helpers.mts";
 
 const helpCommand = new SlashCommandBuilder()
   .setName('help')
@@ -19,6 +19,10 @@ const helpCommand = new SlashCommandBuilder()
       .addSubcommand(command => command
         .setName('stop')
         .setDescription('Termina la conversación con el chatbot si estas todavía en ello.')
+      )
+      .addSubcommand(command => command
+        .setName('status')
+        .setDescription('Muestra el estado de tu conversación actual.')
       )
   })
 
@@ -62,7 +66,7 @@ const startChatbotAction = async (interaction: ChatInputCommandInteraction) => {
 
   const user = User.getUser(interaction.user)
   if (user.isInChat()) {
-    await interaction.editReply({ 
+    await interaction.editReply({
       content: `Estas actualmente en una sesión de chat. Usa \`help endchat\` para cerrar el chat.`
       , components: []
     })
@@ -80,7 +84,7 @@ const startChatbotAction = async (interaction: ChatInputCommandInteraction) => {
   //await thread.members.add(user.getID())
   await interaction.deleteReply()
   let AIMessage = await thread.send(`Hola <@${user.getID()}>, ¿en que puedo ayudarte hoy?`)
-  user.lastIAMessage = AIMessage
+  user.AIChat.lastIAMessage = AIMessage
   while (user.isInChat()) {
     try {
       const collected = await thread.awaitMessages({
@@ -88,51 +92,63 @@ const startChatbotAction = async (interaction: ChatInputCommandInteraction) => {
         max: 1,
         time: 15 * 60 * 1000
       })
-
-      if(collected.size === 0){
+      
+      if (!threadStillExists(user.AIChat.lastIAMessage.guild, user.AIChat.lastIAMessage.channelId)) {
+        user.endChat()
+        return;
+      }
+      if (collected.size === 0) {
+        
         await AIMessage.edit(AIMessage.content + '\n - **Se ha cerrado la conversación por inactividad**')
         user.endChat()
         return
       }
       const userResponse = collected.first();
+      
 
       // en caso de estar actualizando el cache avisa al usuario.
-      if(Bot.isUpdatingCache()){
+      if (Bot.isUpdatingCache()) {
         let response = await userResponse.reply(`<@${userResponse.author.id}> actualmente estoy actualizando mi conocimiento sobre la documentación de facturascripts, puedo tardar hasta 5-10 minutos. Te responderé en cuanto termine ||(borraré este mensaje cuando acabe)||.`)
         await Bot.awaitCacheLoading()
         await response.delete()
       }
+      
       await (userResponse.channel as TextChannel).sendTyping()
 
       let usermsg = await replaceMentionsWithUsernames(userResponse.content, userResponse.guild)
       let generatedMessage = await user.sendMessage(`[Nombre de usuario: <@${userResponse.author.username}>]:` + usermsg)
       generatedMessage = await replaceUsernamesWithMentions(generatedMessage, userResponse.guild)
 
-      if(generatedMessage.includes('$$END_CHAT$$')) {
+      if (generatedMessage.includes('$$END_CHAT$$')) {
         let msg = generatedMessage.replace('$$END_CHAT$$', '').trimEnd()
-        if(msg.trim() !== ""){
+        if (msg.trim() !== "") {
           let sepparatedMessages = splitFromJumpLines(msg, 2000)
           let last = userResponse
           for (const mensaje of sepparatedMessages) {
-            last = await last.reply({content: mensaje})
+            last = await last.reply({ content: mensaje })
           }
-          user.lastIAMessage = last
+          user.AIChat.lastIAMessage = last
         }
-        await user.lastIAMessage.reply('- **Se ha cerrado la conversación (dada por finalizada)**')
+        await user.AIChat.lastIAMessage.reply('- **Se ha cerrado la conversación (dada por finalizada)**')
         user.endChat()
         return
-      }else{
+      } else {
         let sepparatedMessages = splitFromJumpLines(generatedMessage, 2000)
         let last = userResponse
         for (const mensaje of sepparatedMessages) {
-          last = await last.reply({content: mensaje})
+          last = await last.reply({ content: mensaje })
         }
-        user.lastIAMessage = last
+        user.AIChat.lastIAMessage = last
       }
 
-      
+
 
     } catch (err) {
+      if (err.code === 'ChannelNotCached') {
+        user.endChat()
+        return
+      }
+      
       AIMessage.reply('- **Se ha cerrado la conversación (error interno del bot)**')
       console.error(err)
       user.endChat()
@@ -141,26 +157,56 @@ const startChatbotAction = async (interaction: ChatInputCommandInteraction) => {
   }
 }
 
-async function stopChatbotAction(interaction: ChatInputCommandInteraction)
-{
+async function statusChatbotAction(interaction: ChatInputCommandInteraction) {
   let user = User.getUser(interaction.user)
-  if(user.isInChat()){
-    user.lastIAMessage = await user.lastIAMessage.reply('- **Se ha cerrado la conversación (cerrada por el usuario)**')
-    user.endChat()
-    // await interaction.reply({
-    //   content: "Se ha cerrado la conversación correctamente.", 
-    //   flags: MessageFlags.Ephemeral
-    // })
-  }else{
+  let embed = new EmbedBuilder()
+  if (user.isInChat()) {
+    if (!threadStillExists(user.AIChat.lastIAMessage.guild, user.AIChat.lastIAMessage.channelId)) {
+      embed.setColor(0x007BFF)
+        .setTitle('🔴 Actualmente en chat')
+        .setDescription('Estas en un chat pero se ha borrado.\n Ejecuta `/help chatbot stop` para cerrarlo.')
+    } else {
+      embed.setColor(0x007BFF)
+        .setTitle('🟢 Actualmente en chat')
+        .setDescription(`Estas en un chat en <#${user.AIChat.lastIAMessage.channelId}>.`)
+    }
+  } else {
+    embed.setColor(0x007BFF)
+        .setTitle('🔴 No estás en ningún chat')
+        .setDescription('Ejecuta `/help chatbot start` para entrar en un chat si quieres.')
+  }
+  await interaction.reply({
+    embeds: [embed],
+    //flags: MessageFlags.Ephemeral
+  })
+}
+
+async function stopChatbotAction(interaction: ChatInputCommandInteraction) {
+  let user = User.getUser(interaction.user)
+  if (user.isInChat()) {
+    if (!threadStillExists(user.AIChat.lastIAMessage.guild, user.AIChat.lastIAMessage.channelId)) {
+      user.endChat()
+      await interaction.reply({
+        content: "Se ha cerrado correctamente la conversación.",
+        flags: MessageFlags.Ephemeral
+      })
+    }else{
+      user.endChat()
+      user.AIChat.lastIAMessage = await user.AIChat.lastIAMessage.reply('- **Se ha cerrado la conversación (cerrada por el usuario)**')
+      await interaction.reply({
+        content: "✅",
+        flags: MessageFlags.Ephemeral
+      })
+    }
+  } else {
     await interaction.reply({
-      content: "No hay ninguna conversación abierta.", 
+      content: "No hay ninguna conversación abierta.",
       flags: MessageFlags.Ephemeral
     })
   }
 }
 
-async function helpAction(interaction: ChatInputCommandInteraction)
-{
+async function helpAction(interaction: ChatInputCommandInteraction) {
   if (interaction.options.getSubcommandGroup() === 'chatbot') {
     switch (interaction.options.getSubcommand()) {
       case 'start':
@@ -168,6 +214,9 @@ async function helpAction(interaction: ChatInputCommandInteraction)
         break;
       case 'stop':
         await stopChatbotAction(interaction)
+        break;
+      case 'status':
+        await statusChatbotAction(interaction)
         break;
       default:
         break;
